backend/requirements.txt
# CreAItion Backend Requirements - Simplified

# FastAPI and Web Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.2
python-dotenv>=1.0.0

# Database
supabase>=2.0.2

# AI/ML Dependencies
google-genai>=0.3.2
Pillow>=10.1.0